+++ pyobjc-framework-ClassKit/PyObjCTest/test_clsdefines.py
@@ -27,3 +27,7 @@ class TestCLSActivity(TestCase):
         self.assertIsInstance(ClassKit.CLSPredicateKeyPathUniversalLinkURL, str)
         self.assertIsInstance(ClassKit.CLSPredicateKeyPathTopic, str)
         self.assertIsInstance(ClassKit.CLSPredicateKeyPathParent, str)
+
+    @min_os_level("12.0")
+    def test_constants12_0(self):
+        self.assertIsInstance(ClassKit.CLSErrorSuccessfulObjectsKey, str)