+++ pyobjc-framework-CoreWLAN/PyObjCTest/test_corewlantypes.py
@@ -45,6 +45,7 @@ class TestCoreWLANTypes(TestCase):
         self.assertEqual(CoreWLAN.kCWPHYMode11g, 3)
         self.assertEqual(CoreWLAN.kCWPHYMode11n, 4)
         self.assertEqual(CoreWLAN.kCWPHYMode11ac, 5)
+        self.assertEqual(CoreWLAN.kCWPHYMode11ax, 6)
         self.assertEqual(CoreWLAN.kCWInterfaceModeNone, 0)
         self.assertEqual(CoreWLAN.kCWInterfaceModeStation, 1)
         self.assertEqual(CoreWLAN.kCWInterfaceModeIBSS, 2)