+++ pyobjc-framework-ReplayKit/PyObjCTest/test_rpscreenrecorder.py
@@ -1,4 +1,4 @@
-from PyObjCTools.TestSupport import TestCase
+from PyObjCTools.TestSupport import TestCase, min_os_level
 
 import ReplayKit
 import objc
@@ -37,6 +37,22 @@ class TestRPScreenRecorder(TestCase):
         self.assertResultIsBOOL(ReplayKit.RPScreenRecorder.isRecording)
         self.assertResultIsBOOL(ReplayKit.RPScreenRecorder.isMicrophoneEnabled)
         self.assertResultIsBOOL(ReplayKit.RPScreenRecorder.isCameraEnabled)
+
+    @min_os_level("12.0")
+    def test_methods12_0(self):
+        self.assertArgIsBlock(
+            ReplayKit.RPScreenRecorder.startClipBufferingWithCompletionHandler_,
+            0,
+            b"v@",
+        )
+        self.assertArgIsBlock(
+            ReplayKit.RPScreenRecorder.stopClipBufferingWithCompletionHandler_, 0, b"v@"
+        )
+        self.assertArgIsBlock(
+            ReplayKit.RPScreenRecorder.exportClipToURL_duration_completionHandler_,
+            2,
+            b"v@",
+        )
 
     def test_protocols(self):
         objc.protocolNamed("RPScreenRecorderDelegate")